+++ changelog.md
@@ -4,6 +4,7 @@
 * Added helpers functions to Bitmap and Icons
 * Added the `Monitor` struct to query monitor and screen information
 * Added `center` to the window builder to center a window on screen
+* Fixed the text input refresh after calling `set_password_char`
 
 1.0.8